+++ R/mirai.R
@@ -1367,7 +1367,7 @@ print.mirai <- function(x, ...) {
 #'
 print.miraiError <- function(x, ...) {
 
-  cat(strcat("'miraiError' chr ", x), file = stdout())
+  cat(sprintf("'miraiError' chr %s\n", x), file = stdout())
   invisible(x)
 
 }
@@ -1527,8 +1527,8 @@ mk_mirai_error <- function(e) {
   x <- .subset2(e, "call")
   call <- if (length(x)) deparse(x, width.cutoff = 500L, backtick = TRUE, control = NULL, nlines = 1L)
   msg <- if (is.null(call) || call == "eval(expr = ._mirai_.[[\".expr\"]], envir = ._mirai_., enclos = NULL)")
-    sprintf("Error: %s\n", .subset2(e, "message")) else
-      sprintf("Error in %s: %s\n", call, .subset2(e, "message"))
-  cat(msg, file = stderr());
+    sprintf("Error: %s", .subset2(e, "message")) else
+      sprintf("Error in %s: %s", call, .subset2(e, "message"))
+  cat(strcat(msg, "\n"), file = stderr());
   `class<-`(msg, c("miraiError", "errorValue"))
 }